Modify asn1 generated decode/2 functions to return {error,incomplete}
in the case where the argument packet is incomplete.

This is necessary when a packet received from a socket doesn't contain a full
message (in case of large messages). It complements the
feature obtained by compiling the asn1 file using the undec_rest
option : in case of a packet containing more than one message, the
decode functions return the decoded message and the remaining bytes.